test : RecentlyUser, User 서비스부분 테스트코드 작성

+++ src/main/java/com/ll/TeamSteam/domain/user/service/UserService.java
@@ -152,7 +152,9 @@ public class UserService {
         User user = findBySteamId(steamId).orElseThrow();
         Long userId = user.getId();
         UserTag userTag = userTagService.findByUserId(userId);
-        gameTagService.deleteAll(userTag.getGameTag());
+        if(Optional.ofNullable(userTag.getGameTag()).isPresent()) { //존재하면 지움
+            gameTagService.deleteAll(userTag.getGameTag());
+        }
         if(!steamGameLibraryService.findAllByUserId(userId).isEmpty()){
             steamGameLibraryService.deleteByUserId(userId);
         }
